refactor: extract music notation utils to separate module
